Add main.py to run the app and insert data from CSV
This commit adds a new file, main.py, which contains the main function to run the app. It also includes the functionality to insert data from a CSV file to the database. The main function accepts command line arguments to determine whether to run the app or insert data. If the command is "run", it calls the run_app() function from app.py. If the command is "insert", it calls the insert_data_from_csv() function from data_insertion.py, passing the CSV file path as an argument.